PF372 ajoute les tests de DossiersController
